TIP-221: Move VariantGroupAttributesResolver
